apps/backend/pyproject.toml
[project]
name = "deadpartymedia-backend"
version = "0.1.0"
description = "Django backend for Dead Party Media blog platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.0",
    "djangorestframework>=3.15.0",
    "django-allauth>=65.0.0",
    "django-jazzmin>=3.0.0",
    "django-cors-headers>=4.4.0",
    "django-filter>=24.3.0",
    "django-storages>=1.14.4",
    "django-ckeditor>=6.7.1",
    "django-anymail[resend]>=11.0.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.8",
    "celery>=5.4.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.7.0",
    "python-decouple>=3.8",
    "boto3>=1.35.0",
    "pillow>=10.4.0",
    "twilio>=9.3.0",
    "django-ratelimit>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.django]
settings_module = "backend.settings.dev"

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=4.4.0",
    "black>=24.8.0",
    "isort>=5.13.0",
    "flake8>=7.1.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "factory-boy>=3.3.0",
]